refactor vuex store
