+++ client/src/services/UserService.js
@@ -1,5 +1,5 @@
 import axios from 'axios';
-import store from '../store.js';
+import store from '../store/store.js';
 import router from '../router.js';
 
 /**
